# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GlobalChat API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile (creates if first-time)
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all registered users (for search / adding contacts)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a specific user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── Groups ──────────────────────────────────────────────────────────────────
...
  /groups/{groupId}/messages:
    get:
      operationId: listGroupMessages
      tags: [messages]
      summary: Fetch message history for a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: before
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageList"
    post:
      operationId: sendGroupMessage
      tags: [messages]
      summary: Send a message to a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /groups/{groupId}/messages/{messageId}:
    delete:
      operationId: deleteGroupMessage
      tags: [messages]
      summary: Delete a message (admin or own message)
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Private / DM messages ────────────────────────────────────────────────────
  /dm/{contactUserId}/messages:
    get:
      operationId: listDmMessages
      tags: [messages]
      summary: Get DM conversation with a contact
      parameters:
        - name: contactUserId
          in: path
          required: true
          schema:
            type: string
        - name: before
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: DM messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageList"
    post:
      operationId: sendDmMessage
      tags: [messages]
      summary: Send a direct message
      parameters:
        - name: contactUserId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent DM
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ─── Contacts ─────────────────────────────────────────────────────────────────
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users with ban status (admin only)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [all, active, banned]
            default: all
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Admin user list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
...
  /admin/messages/{messageId}:
    delete:
      operationId: adminDeleteMessage
      tags: [admin]
      summary: Admin force-delete any message
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        isBanned:
          type: boolean
        banReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, clerkId, username, displayName, email, role, isBanned, createdAt]
    UserUpdate:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
    UserList:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
      required: [users, total]
...
    Message:
      type: object
      properties:
        id:
          type: integer
        content:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
        fileType:
          type: ["string", "null"]
        fileName:
          type: ["string", "null"]
        messageType:
          type: string
          enum: [text, image, file, audio, video]
        senderId:
          type: string
        sender:
          $ref: "#/components/schemas/User"
        groupId:
          type: ["integer", "null"]
        recipientId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, messageType, senderId, sender, createdAt]
    MessageInput:
      type: object
      properties:
        content:
          type: string
        fileUrl:
          type: string
        fileType:
          type: string
        fileName:
          type: string
        messageType:
          type: string
          enum: [text, image, file, audio, video]
          default: text
      required: [messageType]
    MessageList:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        hasMore:
          type: boolean
      required: [messages, hasMore]
...
    AdminUserList:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
      required: [users, total]
...